enh: direct support for config.py in sphinx macro
For sphinx, add support for using a config.py directly in addition
to the existing option for a cmake-configured config.py.in.

Projects may want to avoid a cmake-configured config.py.in file
when they use a service like read-the-docs, which expect to run
sphinx outside of our build environment.