ENH: Add git macros to get basic git project info
Adds git macros to get basic git information such as the
tag or hashcode at the tip of a branch or the current
active branch. The motivation for this is to allow Git
based projects to dynamically encode project version
information in a generated header by leveraging Git's
native capabilities to provide this information.

A generic git macro is provided, blt_git(). This allows
running arbitrary git commands and forms the foundation
for all other macros. All macros can optionally take
a SOURCE_DIR argument to specify the path to the git
repository. If SOURCE_DIR is not specified, CMAKE_CURRENT_SOURCE_DIR
is used internally. In addition, the blt_is_git_repo() macro
is provided to faciliate checking if the source directory is
a valid Git repository or an exported tarball (i.e., no .git folder)

This commit resolves #4.