Fix Issue 221 - Incorrect test for cuda_runtime. (#222)
In line 235 of BLTPrivateMacros.cmake in the definition of the
blt_setup_cuda_target macro, _cuda_index was being used in the test
for cuda_runtime instead of _cuda_runtime_index.  So targets depending
on cuda_runtime where not properly set up. In particular, executables
depending on cuda_runtime which wanted to link with nvcc would get link
error because the wer being linked with the host compiler instead of
nvcc.

Changing the test to use _cuda_runtime_index fixes the link errors
for the LEOS project.

Change test to use _cuda_runtime_index.  This fixed the link