ENH: check if C/C++ & Fortran openmp flags differ
This commit adds a check in SetupOpenMP.cmake to
detect if the C/C++ openmp flags are different
from the corrresponding flags used by the
Fortran compiler. This can happen in build
configurations that mix the Fortran and
C/C++ compilers. For example, using Clang compilers
for C/C++ and the GNU fortran compiler for Fotran.

A cache variable, BLT_OPENMP_FLAGS_DIFFER, is added
which will be set to true if the C/C++ openmp flags
are different from the Fotran openmp flags, or set
to false otherwise.

This commit resolves in part #175.