ENH: fix the fortran openmp flags if needed
This commit adds a function to fix the fortran openmp flags
for a given target if needed. The function queries the LINK_FLAGS
target property and if C/C++ OpenMP flags are different from the
Fortran flags, the function replaces the C/C++ flags with the
Fortran flags.

The function is called from within blt_add_executable() on Fortran
targets.

Resolves #175.