ENH: add NVIDIA SDK path to the CUDA link flags
This commit adds the path to the NVDIA SDK libraries to the
CMAKE_CUDA_LINK_FLAGS. Moreover, it ensures that these
link flags are specified to the registered CUDA library that
is used with BLT.

This allows applications to link additional libraries that are supplied
with the NVIDIA SDK by listing them in the list of dependencies,
e.g., nvToolsExt, cublas, etc. In addition, CMAKE_CUDA_LINK_FLAGS
may also be specified in a host-config or at the command line and
those will now be correctly propagated and associated with
BLT's registered CUDA library.