fix linking issue in hip 3.5
This fixes the rest of #366 so that the smoke test links with hip 3.5.  I'm going to feed this bug up to AMD as well, since this change should not be necessary.

The `STR_LENGTH` should also be constexpr, or at least __host__ __device__, since the current version is only valid because of over-eager constant propagation by the hip compiler, but I'm leaving that alone for now.