feat: yapf: expose ENABLE_YAPF variable
This commit adds an ENABLE_YAPF variable intended for eventual use in
supporting Python source beautification via
[YAPF](https://github.com/google/yapf) -- Yet Another Python
Formatter.

YAPF was chosen over Autopep8 and Black because YAPF is more
configurable than both and Spack uses it. For the purposes of this
branch, adding Autopep8 or Black support is considered out of scope.