Use find_library to search for HIP rutime libraries
This addresses PR comments and will hopefully be more robust.
The nvidia route isn't well tested. I know it will build and run the
BLT smoke tests but it can't build and run RAJA, but I think that might
be due to RAJA assuming that the HIP backend will have rocPRIM.
The NVIDIA path was broken prior to this PR because it used
find_package(cudatoolkit) which needed to be find_packed(CUDAToolkit).
This makes me think the HIP_PLATFORM=nvidia route is never exercised
and maybe shouldn't be supported at all.